Fix accidental side effect of setlabel.
Change behavior when setlabel is called to apply a label to an address which is
already the default address of the same label. Instead of deleting the label
address and causing the next getlabeladdress call to return a brand new
address, do nothing and keep everything the same.

This way calling setlabel to an apply a label to an address which already has
that label will simply be a no-op and not cause unintended side effects.