RPC/Mining: prioritisetransaction: Restore priority_delta argument to not gratuitously break compatibility between implementations
Other implementations (as well as all current versions of Core) expect the fee-delta to be the third parameter.
Since priority is no longer supported in this implementation, it throws an exception if the priority delta is non-zero.

This partially reverts commit f9b9371c6027905f73a2558d6bcaca8a355c28a6.