Add barebones IPC framework to bitcoin-qt and bitcoind
This adds some IPC functions and classes in src/ipc/ that let bitcoin-qt fork a
bitcoind process and communicate with it over a socketpair.

Only a handful of IPC calls are implemented, but this demonstrates a working
framework supporting bidirectional IPC calls (requests from bitcoin-qt to
bitcoind and notifications from bitcoind to bitcoin-qt).

The IPC protocol used in this implementation is Cap'n Proto
(https://capnproto.org/), but this could be swapped out for another protocol.
Can'n Proto types and libraries are only accessed in ipc/server.cpp and
ipc/client.cpp, and not in any headers or other parts of bitcoin code.