Fix bumpfee rpc "errors" return value
Unfortunately I couldn't figure out how to write a test for this bug. It seems
errors return value is never actually populated unless the bumpfee transaction
is rejected from the mempool or there's a database error in
CWallet::MarkReplaced. Probably some mock functionality should be added to
emulate database errors and mempool rejections for better testing in the
future.