[wallet] Make setlabel idempotent
Prevent setlabel calls on an address that already has the same label, and which
is the already the default "label address" of that label, from generating a
new default address.

This way calling setlabel on an address that already has the specified label is
a no-op.

Change suggested by Russell Yanofsky <russ@yanofsky.org> in
https://github.com/bitcoin/bitcoin/pull/7729#discussion_r92064042