[wallet] Make setlabel idempotent
Prevent setlabel calls an address that already has the same label, and which is
the already the default "label address" of that label, from generating a
default address.

Make calling setlabel on an address that already has the label a no-op.

Bug and fix from Russell Yanofsky <russ@yanofsky.org> in
https://github.com/bitcoin/bitcoin/pull/7729#discussion_r92064042