[policy] Prioritize multi-output group in coin selection
With output groups, the 'lowest larger' item will sometimes contain multiple outputs. In this case, prioritizing spending this (address reuse) output is given, unless it is significantly larger than the target value (2x+). The aim here is to get rid of address reuse, and if the user has enabled -avoidpartialspends, the group will *always* be spent together anyway.