threads: introduce threadutil, move RenameProcess, add thread suffixes
This work is prerequisite to attaching thread names to log lines and deadlock
debug utilities. The `threadutil` unit contains platform-specific emulation of
thread_local storage (which isn't yet available for usage on certain platforms
we care about) and is based heavily on work by theuni.

We'd like to persist, per thread, a name and unique identifier. Because
thread_local is not available reliably on platforms we support,
we fall back to pthreads' setspecific/getspecific interface. If
pthreads is not available, we use a mutex-guarded map.

This commit also moves RenameProcess() out of util, removes the "bitcoin-"
prefix from subprocesses, and adds a numeric suffix to disambiguate between
threads with the same name.