Make checkqueue bench a realistic result
Previously checkqueue bench would essentially just test cross-core
latency with more threads the more threads a host had. Not only did
this make the result worse on hosts with more CPU cores, but it
doesn't mimic how we use the checkqueue in real life, and made
bench_bitcoin almost unusably slow on high-core-count systems,
especially dual-CPU systems.

Instead, we use the same thread count that we use for the real
validation queue, with a 75-microsecond busy-wait on each thread,
which should at least vaguely mimic how we use the checkqueue.