threads: introduce threadutil, refactor thread naming
This work is prerequisite to attaching thread names to log lines and deadlock
debug utilities. threadutil allows setting of an "internal" threadname per
thread on platforms where thread_local is available.

This commit also moves RenameThread() out of util and adds a numeric suffix to
disambiguate between threads with the same name.