Don't assign DoS points when -promiscuousmempool is set
Reset CValidationstate() when a transaction fails policy standard
but not mandatory validation flags when -promiscuousmempool is set

Currently, if -promiscuousmempool is set (only allowed on regtest
and testnet), it is possible for AcceptToMemoryPool() to return
true, but for the CValidationState() to contain DoS points if the
transaction failed validation with policy standard script flags
but then passed with promiscuous flags. This will cause you to ban
the peer who sent the transaction. This commit fixes this behavior
by reseting the validation state in that one situation.