Refactor CheckInputs() to only validate scripts once
Previously CheckInputs() could invoke a script twice to determine if a script
failure was due to standardness rules. However, it only performed this in
single-threaded validation mode, and it directly accessed policy variables to
make this determination (even though the caller was passing in the desired
script flags to use).

The policy vs consensus script flags comparison seems out of touch with current
network deployments; the only MANDATORY script flag is P2SH, even though many
other consensus changes have been deployed since. However, adding further flags
to the MANDATORY set risks making the p2p behavior overly restrictive by
disconnecting unupgraded peers unnecessarily.

Moreover, AcceptToMemoryPoolWorker() already invokes CheckInputs multiple
times, with different script flags, to detect certain kinds of segwit
validation errors. Given that the caller is responsible for passing in the
correct validation flags, it feels like a layer violation for CheckInputs() to
be re-running checks with different script flags by itself.

This patch moves all responsibility for determining which script flags are
mandatory to the caller, so CheckInputs() will now always return CONSENSUS for
any script failures. The caller can rewrite the error code itself, eg if
policy-flags were included.

This patch also changes the ban behavior for peers sending transactions with
invalid scripts; we will no longer disconnect such peers. It was already
trivial for an adversary to mutate a transaction to violate a policy rule,
previously, in order to relay a transaction and avoid a ban, so we were not
relying on the ban behavior previously.

As this results in changes to the reject reasons and codes we can return, many
of the tests have been updated as well.