move-only: make the CChainState interface public
Along with DisconnectResult, and CBlockIndexWorkComparator (in a private
namespace).

The CChainState interface needs to be known to the rest of the system because
many global functions will move to CChainState methods. This is to allow
other parts of the system to be parameterized per chainstate instance
instead of assuming a single global.