Multiprocess build changes
autotools and automake changes to support multiprocess execution.

This adds a new --enable-multiprocess flag, and build configuration code to
detect libraries needed for multiprocess support. The --enable-multiprocess
flag builds new bitcoin-node, bitcoin-wallet, and bitcoin-gui executables,
which are updated in #10102 communicate across processes. But for now they are
functionally equivalent to existing bitcoind and bitcoin-qt executables.