validation: indexing changes for assumeutxo
Adds BackgroundBlockConnected to the validationinterface and uses it in
index maintenance. Ensures that index building will work when part of
the chain is validated asynchronously in the background by a second
chainstate.

This changeset removes the guarantee that indexes will be built
sequentially, as none of the current indexes require this and it is no
longer easy to offer this guarantee when multiple chainstates are in
use.

Within BaseIndex, we only update `m_best_block_index` (which is
essentially the progress marker for how far along the indexing process
is) for chains which we can be certain that all blocks under a given
block on that chain have had indexing performed on them.  In other
words, we will not update `m_best_block_index` during a BlockConnected
event that comes from an active snapshot chain, i.e. a chain for which
some of the blocks underneath Tip() are not indexed.

Once background validation is completed and no background chainstates
are in use, the indexer will happily use BlockConnected events from the
active chain to update `m_best_block_index` as usual.