Change feature_config_args.py not to rely on strange regtest=0 behavior
Update test to simply generate a normal mainnet configuration file instead of
using a crazy setup where a regtest=1 config file using an includeconf in the
[regtest] section includes another config file that specifies regtest=0,
retroactively switching the network to mainnet.

This setup is fragile and only works because InitError it is checking early
enough that none of the ignored [regtest] options the test framework is trying
to pass matter too much (maybe affecting log output at most).