cli: extract connection exception handler, -rpcwait logic
to ConnectAndCallRPC() to be callable for individual connections.

This is needed for RPCs that need to be called and handled sequentially, rather
than alone or in a batch.

For example, when fetching the balances for each loaded wallet, -getinfo will
call RPC listwallets, and then, depending on the result, RCP getbalances.

Best to review this commit with `git show -w`.

Github-Pull: #18594
Rebased-From: 29f2cbdeb7afdde87d108adf80cffad17d112632