qt: Do not call setParent() for WalletModel instances
Firstly, setParent() is redundant because WalletModel instance lifetime
is already managed by WalletController::removeAndDeleteWallet().
Secondly, setParent() calls sendEvent(Parent, QChildEvent) that implies
running in the thread that created the WalletController object, i.e.,
the GUI main thread. That is not the case always, and ASSERT failure is
emitted in the debug mode.