Fix lock stack after "deadlock detected" exception
If the UniqueLock() ctor throws the "potential deadlock detected"
exception, the object is not instantiated, and the ~UniqueLock() dtor
will never be called for it.
These changes fix broken behavior which creates a "ghost" lock in the
lock stack: this lock is pushed to the stack by the UniqueLock() ctor,
but it is never popped from the stack by the ~UniqueLock() dtor.