build: pass _WIN32_WINNT=0x0601 when building libevent for Windows
This enables of the use of AI_* definitions in the Windows headers,
specifically AI_ADDRCONFIG, which fixes an issue with libevent and
ipv6 on Windows.

It also aligns with what we define in configure when building Core.

Github-Pull: #19375
Rebased-From: eb6b73540d1ee7ff5a6874dd0e35f9b30b68e3b8