Move setDirty{BlockIndex,FileInfo} to BlockManager
Move these variables to BlockManager because they are used when updating
the block index on disk.

In the future it might make sense to move more logic responsible for
updating and writing the block index to the BlockManager class and make
these variables private instead of public. See instances of CChainState
reaching into m_blockman.setDirty* for inspiration.