Better error messages for invalid addresses
This commit addresses #20809.

We add an additional 'error' property in the result of 'validateaddress' in case the address is not valid that gives a short description of why the address in invalid. We also change the error message returned by 'getaddressinfo' in case the address is invalid.

Github-Pull: #20832
Rebased-From: 8f0b64fb513e8c6cdd1f115856100a4ef5afe23e