rpc: Add a `-rpcwaittimeout` parameter to limit time spent waiting
Adds a new numeric `-rpcwaittimeout` that can be used to limit the
time we spend waiting on the RPC server to appear. This is used by
downstream projects to provide a bit of slack when `bitcoind`s RPC
interface is not available right away.

Github-Pull: #21056
Rebased-From: a7fcc8eb59fe51473571661316214156fbdbdcae