qt: Add Copy Address Action to Payment Requests
Currently, the only way to copy the address of a payment request is to double-click on the payment request and then click on the copy address button. This commit adds a convenient context menu action to copy the address of a payment request.

Github-Pull: gui#213
Rebased-From: e348d7ea2c0b5d2eb8039dab33b0b9a48655885f