validation: have LoadBlockIndex account for snapshot use
When reconnecting the block index on startup, we have to account for the
fact that assumed-valid blocks (per the snapshot chainstate) do not have
nTx values. We can't exclude blocks from inclusion in
`setBlockIndexCandidates` on this basis or we will trip over an empty
set in an assertion later in startup.

We also have to ensure that blocks past the snapshot base block (i.e.
the end of the assumed-valid region of the chain) are not included in
setBlockIndexCandidates for the background validation chainstate.