validation: have LoadBlockIndex account for snapshot use
When reconnecting the block index on startup, we have to account
for the fact that assumed-valid blocks (per the snapshot chainstate)
do not have nTx values. We can't exclude blocks from inclusion in
`setBlockIndexCandidates` on this basis or we will trip over an
empty set in an assertion later in startup.