validation: Call NotifyBlockTip in CCS::LoadChainTip
Logically, we only needed to call RPCNotifyBlockChange in ACSS because
the tip was updated in LoadChainTip (see the call to LoadChainTip
earlier in ACSS) but no one was informed of this change.

Seemingly it makes more sense for LoadChainTip to perform this, as it
has more context w/re whether or not the tip was actually changed and a
pointer to the new tip.

Use uiInterface.NotifyBlockTip instead of RPCNotifyBlockChange since the
former calls the latter via signals, and is what CS::ActivateBestChain
and CS::InvalidateBlock does.