rpc: Explain active and internal in listdescriptors
The current help text for active and internal in listdescriptors is not
particularly helpful. They require the reader to already know what those
terms mean. This help text is updated to actually explain the
definitions of those words in context of a descriptor wallet.

Github-Pull: #24977
Rebased-From: 4637bbe448ae7370528f40092ce230c32602a6d6