consensus: allow OP_UNVAULT outputs to be behind scripthash
Previously, OP_UNVAULT outputs had to be bare scripts so that the script
interpreter could inspect them during OP_VAULT spend to ensure that the
parameters of the new output were compatible.

Greg Sanders had the idea that if we put the <target-hash> on the
witness stack, we could construct and match against expected
scriptPubKeys for P2WSH and P2TR. This allows the full unvault lifecycle
to live within witness outputs.

Co-authored-by: Greg Sanders <gsanders87@gmail.com>