consensus: implement OP_VAULT and OP_UNVAULT
OP_VAULT adds a native way of doing vaults - that is, encumbering coins
in such a way that they must pass through a delay period before being
successfully spent, aside from a predetermined path (the recovery
path) which they can be swept to at any time during the life of the
vault.

Vaulting
========

To create a vault, coins must be encumbered under a scriptPubKey that
looks like

  <recovery-params> <spend-delay> <trigger-sPK-hash> OP_VAULT

Of course, this can be within a P2TR taptree.

<recovery-params>
-----------------

The `recovery-params` paramater is of the form

  <32-byte-hash>[<arbitrary scriptPubKey>]

where the first component is the hash of the recovery path scriptPubKey:
i.e. the scriptPubKey which the full value of recovered vaults must be
spent to in order to successfully sweep the coins from a vault.

The optional bytes following the recovery path hash dictate the optional
recovery authorization. When this is specified, the script given must be
satisfied when sweeping the vaulted inputs to recovery. Using this
parameter has benefits and drawbacks. The benefits are

  - unrelated inputs and outputs can be attached to a recovery, making
    "exogenous" fee management like ephemeral anchor use unneccessary.

  - vaults with incompatible recovery parameters can be recovered in the
    same transaction, which allows fee managed UTXOs to be shared.

however the drawbacks are

  - increased risk/complexity around the ability to perform recovery
    sweeps.

  - if an attacker obtains the recovery authorization key and all
    information necesary to attempt a recovery (the "location" of the
    vaulted coins) as well as the trigger authorization key, they may
    attempt to both trigger an unvault and pin a recovery. This is
    mitigated by the necessary presence of an ephemeral anchor output
    in every recovery transaction.

In either case, any vault outputs which share <recovery-params> can be
recovered together in the same transaction.

<spend-delay>
-------------

This is a relative timelock a la BIP-68 that specifies how many blocks
(or how much time) must elapse before an OP_UNVAULT output becomes
spendable to the declared unvault target (more on this below).

Vaults with compatible <recovery-params> and <spend-delay> can be
unvaulted together.

<trigger-sPK-hash>
------------------

This is the hash of the scriptPubKey which needs to be satisfied in
order to trigger the unvault process.

OP_UNVAULT
==========

To trigger the start of the unvault process, vaulted outputs must be
spent into a transaction which has a corresponding OP_UNVAULT output of
the form

  <recovery-params> <spend-delay> <target-hash> OP_UNVAULT

where the first two parameters are identical to the input vault and the
last parameter, <target-hash>, is a content hash of the proposed unvault
destination. It is a hash of each target output's scriptPubKey and
amount.

Because triggers are always signed with the trigger authorization key,
these transaction can be always be accompanied by unrelated inputs and
outputs, e.g. fee management coins or multiple incompatible vaults.

Triggers support a special "revault" output, which allows depositing
some balance of the vault back into the OP_VAULT scriptPubKey that is
being withdrawn from.

When spending an OP_VAULT into an OP_UNVAULT trigger output, the witness
stack must look like

  [trigger witness stack ...] <trigger-sPK> <target-hash> <trigger-vout_idx>

The first two parameters are evaluated recursively as script. The
<trigger-vout-idx> parameter indicates which vout is the related
OP_UNVAULT, which allows us to handle script upgradeability for future
witness versions.

Final withdrawal
================

Once an OP_UNVAULT output has been confirmed and the related
<spend-delay> has elapsed, the output can be spent into a transaction
matching the <target-hash> given in the trigger transaction. This final
withdrawal transaction will be responsible for providing its own fees in
whatever manner the withdrawer prefers.

Recovery
========

At any point before final withdrawal, OP_VAULT or OP_UNVAULT outputs can
be swept to the recovery path.

If no optional recovery authorization is used, all inputs to the
recovery transaction must be from recovery-compatible vault outputs.
There can only be two outputs: one transferring the full value of the
vault inputs to the recovery sPK, and one ephemeral anchor output.

If an optional recovery authorization has bene specified, the recovery
transaction can be more freeform. The only requirements are that the
full value of any vault inputs are transferred to a compatible recovery
sPK output, and there must be at least one ephemeral anchor output. But
unrelated inputs/outputs can "ride along" or provide fees, and otherwise
recovery-incompatible vaults can be recovered together.

[NOTE TODO: this has yet to be implemented, but is planned]

Other notes
===========

Use of OP_VAULT is restricted to witness v1 and up (Taproot).

Credit
======

Thanks to Greg Sanders, AJ Towns, and John Moffett for material
suggestions and review that wound up improving this proposal.

- AJ: make recovery authorization optional, revault trigger outputs
- Greg: put trigger-hash onto witness stack to avoid bare OP_UNVAULT
- John Moffett: explicating recursive eval attack and providing a test
  case.