Move block-storage-related logic to ChainstateManager
The idea is to fully separate the notion of which blocks are stored on disk,
and what data is in our block index, from what tip a chainstate might be able
to get to. The snapshot-chainstate (if we have one) can use different validation
rules than the background-validation-chainstate in order to treat blocks that
are assumed to be valid as legitimate for advancing the chain. So, we can use
chainstate-agnostic logic to download and store blocks, and let the chainstates
figure out for themselves when to advance the tip.

This commit is missing validation logic (probably something in
FindMostWorkChain and TryAddBlockIndexCandidate) to do the appropriate
filtering so that the two chainstates are able to proceed according to their
different rules. See comment in validation.cpp.

Note: some of the invariants in CheckBlockIndex are modified, but more work is
needed (ie to move CheckBlockIndex to ChainstateManager, as most of what
CheckBlockIndex is doing is checking the consistency of the block index, which
is outside of Chainstate).