Tighten requirements for adding elements to setBlockIndexCandidates
When using assumeutxo, we only need the background chainstate to consider
blocks that are on the chain leading to the snapshotted block.

Note that this introduces the new invariant that we can only have an assumeutxo
snapshot where the snapshotted blockhash is in our block index. As a followup,
unit tests that mock creating a snapshot ought to be updated to reflect this.

Includes test fixes and improvements by Andrew Chow and Fabian Jahr.