Tighten requirements for adding elements to setBlockIndexCandidates
When using assumeutxo with a snapshot chainstate, we only need the background
chainstate to consider blocks that are on the chain leading to the snapshotted
block.

Note that this introduces the new invariant that we can only have a snapshot
chainstate where the snapshotted blockhash is in our block index. Not all unit
tests were enforcing this, so some are removed and will need to be reworked.
See FIXME in validation_chainstatemanager_tests.