gui: Check for private keys disabled before attempting unlock
Before trying to unlock a wallet, first check if it has private keys
disabled. If so, there is no need to unlock.

Note that such wallets are not expected to occur in typical usage.
However bugs in previous versions allowed such wallets to be created,
and so we need to handle them.

Github-Pull: gui#773
Rebased-From: 517c7f9cba306292e12e166b9dbc6c0838f05b27