wallet: Check for descriptors flag before migration
Previously we would check that there is no LegacySPKM in order to
determine whether a wallet is already a descriptor wallet and doesn't
need to be migrated. However blank legacy wallets will also not have a
LegacySPKM, so we need to be checking for the descriptors flag instead.

Github-Pull: #28976
Rebased-From: b1d2c771d43b802db12768e620588ed179e92b29