common: Add PSBTError enum
Add separate PSBTError enum instead of reusing TransactionError enum for PSBT
operations. The error codes returned by PSBT operations and transaction
broadcast functions mostly do not overlap, so using an unified enum makes it
harder to call any of these functions and know which errors actually need to be
handled.

Define PSBTError in the common library because PSBT functionality is
implemented in the common library and used by both the node (for rawtransaction
RPCs) and the wallet.