test: Make sure that migration test does not rescan on reloading
We want to make sure that all of the transactions are being copied to
the watchonly and solvable wallets as expected. The automatic rescanning
behavior can cause us to pass a test by finding the transaction
on loading rather than having it be copied as expected.

Github-Pull: #28868
Rebased-From: 71cb28ea8cb579ac04cefc47a57557c94080d1af