fixup! ci: Test CMake edge cases
Now flags from depends are followed by the per-configuration flags.
We have to override the default "-O0 -g" flags to achieve the same flags
as in the master branch.

For any other configuration, which does not define specific macros etc,
it will be enough to use the "None" configuration to do not override the
flags from depends.