test: Drop `x` modifier in `fsbridge::fopen` call
The MinGW-w64 toolchain links executables to the old msvcrt C Runtime
Library that does not support the `x` modifier for the _wfopen()
function. It is possible to link to the new UCRT library instead.
However, the latter is available on Windows 10 only, which breaks our
release compatibility with the older Windows versions.