fixup! cmake: Include CTest
The current `add_boost_test` implementation assumes that source files
always have paths relative to a source directory. After the previous
commit, it is no longer the case, because sources might have absolute
paths as well.

This change addresses that issue and rewrites the `add_boost_test`
function to accept a single argument, which is a source file absolute
path.